creating articleImgPicker module as preperation for tests
